Core/requirement.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
mail-parser==3.15.0
pytz==2022.4
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
